requirements.txt
discord.py>=2.6.4
aiosqlite>=0.19.0
python-dotenv>=1.0.0
